Added Yigit's tests
